:construction: Try to remove hook. In order to improve perf, trying to remove dispatcherBefore hook usage. MBO module is quite heavy, and do not need to be loaded on every request.
