+++ src/Traits/UseHooks.php
@@ -48,7 +48,6 @@ trait UseHooks
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionListModules;
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionModuleRegisterHookAfter;
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseDisplayEmptyModuleCategoryExtraMessage;
-    use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionDispatcherBefore;
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionObjectShopUrlUpdateAfter;
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionGeneralPageSave;
     use \PrestaShop\Module\Mbo\Traits\Hooks\UseActionBeforeUpgradeModule;
